refactor sharding strategy
