+++ pkg/controller/v1beta1/trainedmodel/controller.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"github.com/kubeflow/kfserving/pkg/constants"
 	"github.com/kubeflow/kfserving/pkg/controller/v1beta1/trainedmodel/reconcilers/modelconfig"
-	"github.com/kubeflow/kfserving/pkg/controller/v1beta1/trainedmodel/sharding"
+	"github.com/kubeflow/kfserving/pkg/controller/v1beta1/trainedmodel/sharding/memory"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
@@ -66,7 +66,7 @@ func (r *TrainedModelReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error
 		return reconcile.Result{}, err
 	}
 	log.Info("Reconciling TrainedModel", "apiVersion", tm.APIVersion, "trainedmodel", tm.Spec)
-	shardStrategy := sharding.MemoryStrategy{}
+	shardStrategy := memory.MemoryStrategy{}
 	shardId := shardStrategy.GetOrAssignShard(tm)
 	// Use tm's parent InferenceService field to get the model configMap
 	modelConfigName := constants.ModelConfigName(tm.Spec.InferenceService, shardId)
