+++ pkg/controller/v1beta1/trainedmodel/sharding/memory/strategy.go
@@ -1,12 +1,8 @@
-package sharding
+package memory
 
 import (
 	v1beta1api "github.com/kubeflow/kfserving/pkg/apis/serving/v1beta1"
 )
-
-type Strategy interface {
-	GetOrAssignShard(trainedModel *v1beta1api.TrainedModel) int
-}
 
 //TODO MemoryStrategy will be implemented in another PR
 type MemoryStrategy struct {
